requirements.txt
streamlit>=1.28.0
langchain-openai>=0.1.0
langchain-core>=0.1.33
langchain>=0.1.0
langgraph>=0.0.26
python-dotenv>=1.0.0
pydantic>=2.5.0
typing-extensions>=4.8.0

# OCR dependencies
pillow>=10.0.0
opencv-python>=4.8.0
pytesseract>=0.3.10
easyocr>=1.7.0
numpy>=1.24.0
